# ipcache: modify labels map in-place in `resolveLabels`